Added UI changes and added Splash screen
