+++ CampusNavigator/CampusNavigator/CampusNavigator/CampusNavigatorApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct CampusNavigatorApp: App {
     var body: some Scene {
         WindowGroup {
-            LoginView()
+            SplashScreenView()
         }
     }
 }
