+++ CampusNavigator/CampusNavigator/CampusNavigator/Models/SuggetionModel.swift
@@ -1,8 +1,8 @@
-//
-//  SuggetionModel.swift
-//  CampusNavigator
-//
-//  Created by Sasindu 034 on 2025-02-27.
-//
-
 import Foundation
+struct ComplaintSuggestion: Identifiable {
+    let id = UUID()
+    var subject: String
+    var details: String
+    var date: Date
+    var isSuggestion: Bool
+}
